+++ smart-diet-backend/src/users/users.service.ts
@@ -1,6 +1,6 @@
 // src/users/users.service.ts
 import { Injectable, ConflictException } from '@nestjs/common';
-import { PrismaService } from '../prisma.service'; 
+import { PrismaService } from '../prisma/prisma.service'; 
 import { SignupDto } from './dto/signup.dto';
 import * as bcrypt from 'bcrypt'; // 비밀번호 암호화를 위한 라이브러리
